# packages/client/configs/openapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Insurer",
    "description": "Insurer API Documentation",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
      "def-0": {
        "description": "a customer object",
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "title": "Customer"
      },
      "def-1": {
        "description": "a product object",
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "title": "Product"
      },
      "def-2": {
        "description": "a contract object",
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "customerId": {
            "type": "number"
          },
          "productId": {
            "type": "number"
          },
          "startDate": {
            "type": "string"
          },
          "renewalDate": {
            "type": "string"
          },
          "claimDate": {
            "type": "string"
          }
        },
        "title": "Contract"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/customers/": {
      "get": {
        "summary": "list all customers",
        "tags": [
          "customers"
        ],
        "description": "list all customers",
        "responses": {
          "200": {
            "description": "Succesful response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Succesful response",
                  "type": "array",
                  "items": {
                    "description": "a customer object",
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "create a customer",
        "tags": [
          "customers"
        ],
        "description": "create a customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "a customer object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a customer object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "summary": "get a customer by id",
        "tags": [
          "customers"
        ],
        "description": "get a customer by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a customer object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a customer object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "update a customer by id",
        "tags": [
          "customers"
        ],
        "description": "update a customer by id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a customer object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a customer object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete a customer by id",
        "tags": [
          "customers"
        ],
        "description": "delete a customer by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succesful response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Succesful response",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{id}/contracts": {
      "post": {
        "summary": "create a contract for a customer by id",
        "tags": [
          "customers"
        ],
        "description": "create a contract for a customer by id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "number"
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "renewalDate": {
                    "type": "string"
                  },
                  "claimDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a contract object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a contract object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "customerId": {
                      "type": "number"
                    },
                    "productId": {
                      "type": "number"
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "renewalDate": {
                      "type": "string"
                    },
                    "claimDate": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/": {
      "get": {
        "summary": "list all products",
        "tags": [
          "products"
        ],
        "description": "list all products",
        "responses": {
          "200": {
            "description": "a product object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a product object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "create a product",
        "tags": [
          "products"
        ],
        "description": "create a product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "a product object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a product object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "summary": "get a product by id",
        "tags": [
          "products"
        ],
        "description": "get a product by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a product object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a product object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "update a product by id",
        "tags": [
          "products"
        ],
        "description": "update a product by id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a product object",
            "content": {
              "application/json": {
                "schema": {
                  "description": "a product object",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete a product by id",
        "tags": [
          "products"
        ],
        "description": "delete a product by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succesful response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Succesful response",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/email": {
      "post": {
        "summary": "send an email notification",
        "tags": [
          "notifications"
        ],
        "description": "send an email notification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": {
                    "type": "string"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "contractStartDate": {
                    "type": "string"
                  },
                  "customerName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesful response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Succesful response",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "customers",
      "description": "Customer related end-points"
    },
    {
      "name": "products",
      "description": "Product related end-points"
    },
    {
      "name": "notifications",
      "description": "Notification related end-points"
    }
  ],
  "externalDocs": {
    "url": "https://swagger.io",
    "description": "Find more info here"
  }
}